[X86] Add a DAG combine to turn (Y ^ (blend (X ^ Y), zero, C)) into (blend X, Y, C)
This can occur from (Y ^ ((X ^ Y) & M)) if the ((X ^ Y) & M) part
has multiples uses. Otherwise we canonicalize to ((X & M) | (Y & ~M))
which is further turned into a shuffle and then into a blend.

With an extra use, only the AND part is turned into a shuffle.
Leaving xors around it.

This patch detects the case and late fixes it into a
blend to remove the xors. We don't want to remove the
one use restrictuib (Y ^ ((X ^ Y) & M))->((X & M) | (Y & ~M))
transform as we don't have blends with immediate control for
all VTs.

Part of CMPLRLLVM-20657 and CMPLRLLVM-20655.

Change-Id: Ibcaeff9aeee9248b8c252b4ddfcf044aa7def532
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/260182
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Topper, Craig <craig.topper@intel.com>
Reviewed-by: Bozhenov, Nikolai <nikolai.bozhenov@intel.com>