[VPlan][SVA] Synchronize back propagation and forward propagation algorithms.
Back propagation in SVA is triggered for loop header PHIs to refine
the analysis results, and this refining is expected to be in sync
with what's done in initial compute(). This patch synchronizes the
back propagation algorithm to use the same compute() method by
outlining it and making it flexible for refinement process. Some
minor NFC refactoring is also done for the newly outlined compute()
method.

JIRA: CMPLRLLVM-21370

Change-Id: If2cda13aec659bb7a05a8cbb100776c389889623
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/260687
Reviewed-by: Elovikov, Andrei <andrei.elovikov@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Bozhenov, Nikolai <nikolai.bozhenov@intel.com>