[DTRANS] CMPLRLLVM-11241: Prevent Bad pointer manipulation when multiple GEPs are squeezed as one large GEP followed by another byte flattened GEP
The AggressiveInstCombine pass can convert a chain of GEPs that are
accessing a zero element into one GEP with multiple indices, followed by
byte flattened GEP that access the rest the data. This byte flattened GEP
can produce Bad pointer manipulation if it doesn't have any information
related to the local pointer analysis.

This change set adds the pointee information related to the zero element
in the long GEP. Then it uses the pointees information in the byte flattened
GEP to find the type that is being accessed.

Change-Id: I5ae8905727eaca6805a660b6217e9c28756262ef
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/260980
Reviewed-by: Rivera, Axel Y <axel.y.rivera@intel.com>
Tested-by: Rivera, Axel Y <axel.y.rivera@intel.com>