[VPlan] Add a method for cloning scalar-loops.
This method is part of codegen and would be used for cloning original
loop for the purpose of use in the creation of scalar-peel and
scalar-remainder loops.

When codegen of scalar-peel and scalar-remainder instruction is
implemented, the 'cloneLoop' method would be called and the upper and
lower-bounds of the scalar-loops would be adjusted to fit the VPlans.

Change-Id: Ic9dbc54cb241d5896ada7ad02200c6b9ecea8e69
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/261165
Reviewed-by: Kozhukhov, Sergey S <sergey.s.kozhukhov@intel.com>
Tested-by: alloy <lab_icla@iclgerrit.sc.intel.com>
Reviewed-by: Bozhenov, Nikolai <nikolai.bozhenov@intel.com>