Add an interface to AAResults to query with an override of the location size.
In principle, it's possible to do all the extra work on the client side of the
interface. But AAResults::getModRefInfo provides a lot of helpful other
functionality that makes it really easy to check if there are conflicting
memory accesses in a larger portion of code, and duplicating that seems
counterproductive.

Change-Id: I2fd3baf46010a5c6aa8143aa9b2ad07d02bdfb91
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/261180
Reviewed-by: Woodworth, Daniel <daniel.woodworth@intel.com>
Reviewed-by: Chawla, Pankaj <pankaj.chawla@intel.com>
Reviewed-by: Creech, Timothy M <timothy.m.creech@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Bozhenov, Nikolai <nikolai.bozhenov@intel.com>