CMPLRLLVM-11241: Collect the data for fields which are arrays and have constant entries (Part 1)
This change set is for collecting the structures which can have
fields that are arrays of integers and may have constant entries.
For example:

   %class.TestClass = type <{i32, [4 x i32]}>

   define void @foo(%class.TestClass* %0, i32 %var) {
     %tmp1 = getelementptr inbounds %class.TestClass, %class.TestClass* %0,
             i64 0, i32 1
     %tmp2 = getelementptr inbounds [4 x i32], [4 x i32]* %tmp1, i64 0, i32 0
     store i32 1, i32* %tmp2
     ret void
   }

The field 1 in %class.TestClass is an array of integers, and entry 0 is
storing i32 1. Therefore this structure will be collected and the
information for this field with the constant will be created.

Change-Id: I19162e2a75ef7c11c711a2cc2ad8008b0af87b76
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/261197
Reviewed-by: Rivera, Axel Y <axel.y.rivera@intel.com>
Tested-by: Rivera, Axel Y <axel.y.rivera@intel.com>