[X86] Improve cost modeling for vector multiply by a uniform power of 2 or power of 2 +/- 1
Multiplying by a power of 2 is the same as shift left by log2 of
the power 2. Normaly InstCombine has already done this transform,
but HIR vectorizer sees it as a Mul since SCEV reverses the Shl.

When multiplying by a power of 2 +/- 1 the backend will prefer
to use a shift left and add/sub if there is no native multiply
available.

Change-Id: I099247fb607f0c83d9094e4ad8e9dcd90bc897d9
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/261172
Reviewed-by: Topper, Craig <craig.topper@intel.com>
Tested-by: Topper, Craig <craig.topper@intel.com>