[1/2][VPlan][HIR] Propagate liveness property of Symbase to PHI nodes.
VPlan's HIR decomposer emits PHIs to blend ambiguous VPValues that
update the same Symbase. This patch fixes decomposer to propagate
liveness properties of such Symbases to the PHIs that are created
for it. If the Symbase is found to be live-out then we create a
VPExternalUse for it, and add the PHI as an operand.

JIRA : CMPLRLLVM-10836

Change-Id: Ic7e533561509469164889069791f8b447e4fb166
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/261689
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Karthik Senthil <karthik.senthil@intel.com>
Reviewed-by: Bozhenov, Nikolai <nikolai.bozhenov@intel.com>