[2/2][VPlan][HIR] Fix VPExternalUses that have multiple live-out operands.
During VPlan's HIR decomposition a VPExternalUse is created to
uniquely represent a single live-out temp/symbase. This implies
that in the decomposed VPlan CFG we expect to see a single
VPInstruction that is live-out for given temp/symbase and this
VPInstruction should be used as operand of corresponding
VPExternalUse. This patch adds a post-processing step in HIR
decomposer to fix ambiguous VPExternalUses that have multiple
operands by making using post-domination information.

Change-Id: I4d8bfc350b69d922cfbe5ab2061b7dc33f951d0c
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/261690
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Karthik Senthil <karthik.senthil@intel.com>
Reviewed-by: Bozhenov, Nikolai <nikolai.bozhenov@intel.com>