CMPLRLLVM-21424: Add the changes 23cd70d from the community
The pulldown from 07/07/2020 included the changes D81296
from community, but this change set has an issue with
PDB debug information and parallel for loop. The fix is
in the following changes: 23cd70d.

This change set is adding those changes from the
community to fix the out of bound access.

Change-Id: I3da8bd19a59a7d52c854c94811fa26e48d05bddd
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/261781
Reviewed-by: Battini, Prasad V <prasad.v.battini@intel.com>
Reviewed-by: Cox, Robert <robert.cox@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Bozhenov, Nikolai <nikolai.bozhenov@intel.com>