[VPlan] Fix last value computation for pre-increment live-out inductions.
If induction's live-out value is computed before increment then
finalization should be done using TripCount - 1 iterations. This
patch enables support for this scenario by updating VPLoopEntities
framework to recognize the pattern before VPInductionFinal
instruction is created and represent this info in the instruction
explicitly. This patch also removes the use of VPLoopEntities in
codegen of VPInductionFinal instruction.

Change-Id: Iad1585462faf46d138acb20e76aeeef4d939a5f6
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/261997
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Karthik Senthil <karthik.senthil@intel.com>
Reviewed-by: Bozhenov, Nikolai <nikolai.bozhenov@intel.com>