[DTRANS] CMPLRLLVM-11241: Collect the data for fields which are arrays and have constant entries (Part 2)
This change set extends the data collection for fields that are arrays with
constant entries. The new changes support when the data is being accessed
through one GEP with multiple indices. The analysis process was added, which
removes candidates if there is something wrong with nested structures and/or
related types. Finally, the constant information collected will be written
into the DTransAnalysis.

Change-Id: Iae27017d15e5937c4d544362db6073ac96f1b2fa
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/262542
Reviewed-by: Battini, Prasad V <prasad.v.battini@intel.com>
Reviewed-by: Cox, Robert <robert.cox@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Romanov, Vlad <vlad.romanov@intel.com>