[NFC][VPlan] Remove bunch of INTEL_CUSTOMIZAION markers in headers
Files affected: IntelVPlanValue.h, IntelVPlan.h. Originally the markup was used
to highlight differences from what has already been open-sourced, but that the
upstreaming has really stopped long time ago. No point in maintaining this now.

Originally, I wanted to preserve INTEL_CUSTOMIZATION around HIR-specific code
and was somewhat able to do that for IntelVPlan.h, but IntelVPlanValue.h is a
mess. Originally, it was supposed that *all* HIR-specifics would be isolated
into a "HIRSpecifics HIR;" member without creating *any* HIR-only interfaces in
the main data structures. Sometime in the past that was violated in the
IntelVPlanValue.h file and lots of interfaces leaked. As such, if we'll ever
start upstreaming VPlan again, these would be better be fixed. For now, just
remove the INTEL_CUSTOMIZATION code because properly guarding HIR specifics in
this file is too troublesome.

Change-Id: I0dd5b77706958d6696322517d8821930e79c9328
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/262477
Reviewed-by: Elovikov, Andrei <andrei.elovikov@intel.com>
Tested-by: Elovikov, Andrei <andrei.elovikov@intel.com>