[VPlan][1/2] Support for implicit unconditional last privates.
The implicit unconditional last privates are the live-out values assigned
in the loop body not under any condition.

This patch implements last privates recognition.
We simply go through all VPInstructions in the loop, find
live-outs that are not phi-nodes and not already known loop entity,
and register them as a loop last-private.

Change-Id: I51be60ebe0a9ea2623da25e8aeced7ce477e0b64
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/262517
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Kelarev, Ivan <ivan.kelarev@intel.com>
Reviewed-by: Elovikov, Andrei <andrei.elovikov@intel.com>