CMPLRLLVM-21632: InstCombine is simplifying the CalledOperand of CallBase instructions by eliminating the use of BitCastOperators.  But is not eliminating those BitCastOperators themselves, because it currently only eliminates Instructions. They remain around in the Use lists of Values. This is particularly bothersome when these BitCastOperators include references to Functions, because a Functions can appear to have multiple Uses, when in fact it has only one active Use. This biases our inlining heuristics, which are intend to give a strong preference to inlining single callsite Functions.
NOTE: I have a temporary commented line in InstCombineInternal.h, which
enables this optimization only for Fortran. That is becuase turning it on
for C/C++ causes a 3-4% regression in 511.povray on zperf. There are a lot
more single callsite functions being inlined, and apparently that isn't good
in this case. But we need this change now for 503.bwaves, which is all
Fortran. I will leave CMPLRLLVM-21632 open until I have solved how to get this
optimization enabled without regressing any CPU2017 C/C++ benchmarks.

Change-Id: I17cef76df2e03d9dfaf646eab62ef820e9bb0e76
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/262768
Reviewed-by: Battini, Prasad V <prasad.v.battini@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Lin Jr, Chang-Sun <chang-sun.lin.jr@intel.com>
Reviewed-by: Romanov, Vlad <vlad.romanov@intel.com>