[NFC] CMPLRLLVM-20532: Clean up for Whole program analysis
This change set is a clean up for the whole program analysis before
including the changes related to aliases. The changes are simple,
the Module and the lambda function to collect library functions
are now private fields of the WholeProgramInfo class. The
call graph analysis was removed since it wasn't used.

Change-Id: Id36102e1f38f050b7e55149459a3b17501b9b71d
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/263033
Reviewed-by: Cox, Robert <robert.cox@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Battini, Prasad V <prasad.v.battini@intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Romanov, Vlad <vlad.romanov@intel.com>