[1/3][VPlan] Remove splitLoopsPreheader and splitLoopsExit from construction pipeline.
splitLoopsPreheader and splitLoopsExit are artifacts from the
deprecated hierarchical CFG construction pipeline in VPlan. This
patch removes these transforms for today's flattened VPlan CFG
construction pipeline. Accordingly simplifyCFG is renamed to
doLoopMassaging and associated switches are removed. This is the
first patch in the series for the task to refactor VPlan's
construction pipeline.

Change-Id: I4be44c32531ae7edbbca1b087fa036514ef90335
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/263758
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Karthik Senthil <karthik.senthil@intel.com>
Reviewed-by: Bozhenov, Nikolai <nikolai.bozhenov@intel.com>