[LLD] CMPLRLLVM-21069: Use BFD as the default linker when handling GNU LTO objects
When LLD finds a GNU LTO object, it calls G++ to generate a temporary ELF object
and this object will be used in the final linking process. If LLD is set as the
default linker in the system, this means that G++ will use LLD as linker and not
BFD. The GNU compiler will never be called and the temporary ELF object won't be
be generated, causing missing symbols in the final symbols resolution step.

This change set adds BFD as default linker for the G++ command when LLD is
handing GNU LTO objects. Even if the system is configured to use LLD, it will
use BFD to call the GNU compiler and produce the temporary ELF object.

Change-Id: I015a14c0556f9f6372f7c78542e04384915ae227
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/263261
Reviewed-by: Cox, Robert <robert.cox@intel.com>
Reviewed-by: Chrulski, Christopher M <christopher.m.chrulski@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Romanov, Vlad <vlad.romanov@intel.com>