Fix AP-Int implementation to properly layout non-powers-of-two
The community changed code a bit that resulted in the layout of an
ap-int size to be calculated as too small (in the event that it wasn't a
power of two) because the function to do so rounds down. This caused
other layout problems later.

This patch changes the calculation in this case to round-up in this
situation.

_ExtInt doesn't have this problem, since it is always treated as the
rounded-up size (rather than reporting bitwidth).  Doing so for APInt
would be time consuming, and likely not worth it since it is a
deprecated feature.

Change-Id: I6c74c47922f4856300bd998fa2c2ae4a75d50ce9
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/263432
Reviewed-by: Keane, Erich <erich.keane@intel.com>
Tested-by: Keane, Erich <erich.keane@intel.com>