[Paropt] CMPLRLLVM-21691: remove aritificial branch under -fiopenmp-simd
-fiopenmp-simd sets the paropt mode to "11", turning on ParPrepare, ParTrans,
and OmpVec bits, but does not turn on OmpPar bit.
At Paropt Prepare routine addBranchToEndDirective adds an aritifial branch to
the end of OMP loop. At Paropt Transform removeBranchesFromBeginToEndDirective
removes the artificial branch, but it was done only when the OmpPar bit is set.
The fix makes sure the conditions to create and remove the artificial branch
match.

Change-Id: I6027ada6c97798f82d68d2a1b77439087cf6cb5d
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/263610
Reviewed-by: Gaba, Abhinav <abhinav.gaba@intel.com>
Reviewed-by: Zakharin, Vyacheslav P <vyacheslav.p.zakharin@intel.com>
Tested-by: Su, Ernesto <ernesto.su@intel.com>
Reviewed-by: Bozhenov, Nikolai <nikolai.bozhenov@intel.com>