[FPGA] Add callback to update kernel-event map for FPGA emu
During serialize execution of instances of the same kernel in OOO queue,
one kernel-event map is used to track the status of one kernel execution.
But the tracker event recorded in this map can be released before another
kernel execution checks the status of this event. This may lead to seg
fault.

This fix adds one callback for the tracker event. When its status is
updated to CL_COMPLETE, it's removed from the kernel-event map as it will
not block next kernel execution.

Change-Id: Id3afd35f88f994cb232be781924c0864e4cdcf75
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/261501
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Wenju He <wenju.he@intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Romanov, Vlad <vlad.romanov@intel.com>