[traceback][clang] Support traceback in clang
JIRA:CMPLRLLVM-11793
Add an option "-traceback" for the clang driver, when the driver finds
the option in the command line, it will pass "-traceback" to the
compiler and request the front end to generate the minimal debug
information for traceback. Note that "-traceback" should be orthogonal
to Dwarf/CodeView debug information.

Change-Id: I7de345bbab5e7ba309fdc5d7bb347e3215766e0b
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/264151
Reviewed-by: Kan ShengChen <shengchen.kan@intel.com>
Tested-by: Kan ShengChen <shengchen.kan@intel.com>