[Paropt] Avoid creating dynamic allocas for simd directives
This patch changes placement of alloca instructions that are created for
private variables for simd directives. These instructions are inserted
either into the function's entry block if simd directive does not have
parent region that needs outlining or into the entry block of the parent
region that will be outlined. This is needed to avoid creating dynamic
allocas that can prevent some transformations on functions with simd
directives (for example inlining).

Change-Id: I5e41a86f9d1f5b6720a0fd4213ab5f622b5e89ab
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/264389
Reviewed-by: Dmitriev, Serguei N <serguei.n.dmitriev@intel.com>
Tested-by: Dmitriev, Serguei N <serguei.n.dmitriev@intel.com>