[traceback] Add a debug handler to traverse and write debug info
JIRA:CMPLRLLVM-21785
The debug handler traverses the machine functions and instructions to
collect the debug information of traceback and store them into a data
structure called TraceModule when the module has the flag of traceback,
which is set by the front end. The TraceModule is responsible for
maintaining the debug information in a compact and easy-to-debug format
and writing them into the file.

Dependent patches:
git-amr-2.devtools.intel.com/gerrit/#/c/264180
git-amr-2.devtools.intel.com/gerrit/#/c/264223

Change-Id: Ib01690ea901a9cd956ffd412a974abde5b6504d3
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/264236
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: xiangzhang <xiang1.zhang@intel.com>
Reviewed-by: Bozhenov, Nikolai <nikolai.bozhenov@intel.com>