[DAE][SYCL] Emit MD instead of updating integration header (#2258)
Removed DAE from standard optimization pipeline

We cannot run DAE in ESIMD cntext since the pointers to SPIR kernel
functions are saved in !genx.kernels metadata. Because of that I
removed DAE pass from the standard pipeline and put it in the place
that is specific for SYCL passes and guarded it under a new
temporary option: '-fenable-sycl-dae'. This option is only used until
other parts of the solution are not in place.

This is a cherry-pick of intel/llvm PR #2258:
https://github.com/intel/llvm/pull/2258

It's a part of the implementation of the following feature:
https://jira.devtools.intel.com/browse/CMPLRLLVM-21326

Change-Id: Ifce6ba26ff5138e696227c20e14ff1dd8b176854
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/265858
Reviewed-by: Semenov, Sergey <sergey.semenov@intel.com>
Tested-by: Semenov, Sergey <sergey.semenov@intel.com>
Reviewed-by: Sotkin, Alexey <alexey.sotkin@intel.com>