[traceback][objdump] Dump the debug info of traceback
JIRA:CMPLRLLVM-21786
This patch adds the support for dump tool to interpret the .trace
section. Users can get the meanings of all bytes by using command
```
llvm-objdump --traceback <file>
```
It shows the corresponding tag/attribute of the bytes, prints
the ascii string if the bytes represent a string, displays the
target address if the bytes are the placeholder for a relocation,
errors the unknown byte, incorrect section alignment and warns the
non-optimal encodings. The goal is to interpret as much information
as possible, so the dump tool doesn't check the completeness of the
debug information.

Dependent patches:
git-amr-2.devtools.intel.com/gerrit/#/c/264180
git-amr-2.devtools.intel.com/gerrit/#/c/264236

Change-Id: Ia311c6faf2afe5fa3a992f07dc04fab4e399a9df
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/264386
Reviewed-by: Kan ShengChen <shengchen.kan@intel.com>
Tested-by: Kan ShengChen <shengchen.kan@intel.com>