[2/2][VPlan] Do not widen uniform calls with no side-effects.
Currently all uniform calls are serialized in VPlan LLVM-IR
vectorizer, and this decision is made in CallVecDecisions analysis.
This patch extends the analysis to recognize uniform calls without
any side-effects and record that such calls should not be widened
in outgoing code. VPlan codegen then emits a single scalar call
(predicated with an all zero-check, if needed).

JIRA : CMPLRLLVM-19204

Change-Id: Idcead3cce9438b9738711010d94452cc87c4bb08
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/264771
Tested-by: xmain_checkin_sanity <lab_icls@iclgerrit.sc.intel.com>
Reviewed-by: Karthik Senthil <karthik.senthil@intel.com>
Reviewed-by: Bozhenov, Nikolai <nikolai.bozhenov@intel.com>