CMPLRLLVM-20823: Prevent infinite loop in InstCombine when converting from SExt to Select instructions
InstCombine runs until there is no more instructions to simplify.
In one iteration, it can convert the following expression:

  %cmp = icmp eq i32 %tmp1, -1
  %tmp2 = sext i1 %cmp to i32

Into this Select instruction:

  %tmp2 = select i1 %cmp, i32 -1, i32 0

The issue is that the Select instruction back into the combination of
Icmp and SExt, which will be converted to a Select in the following
iteration. This will produce an infinite loop inside InstCombine.

This change set introduces a simple lock mechanism. Once the instruction
is reverted from Icmp and SExt, it can't be converted as Select again.

Change-Id: Iff092edfdc8528d666829668a519037cdeecc8d3
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/265305
Reviewed-by: Lin Jr, Chang-Sun <chang-sun.lin.jr@intel.com>
Reviewed-by: Cox, Robert <robert.cox@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Bozhenov, Nikolai <nikolai.bozhenov@intel.com>