[SYCL] Add runtime support for device code argument elimination
Add support for the byte array type properties. Register the new kernel
parameter optimization properties while adding binary images in the
program manager. This information is then used while setting kernel
arguments to skip over the eliminated ones.

Signed-off-by: Sergey Semenov <sergey.semenov@intel.com>

This is a cherry-pick of intel/llvm PR #2315:
https://github.com/intel/llvm/pull/2315

It's a part of the implementation of the following feature:
https://jira.devtools.intel.com/browse/CMPLRLLVM-21326

Change-Id: I2e513740a921a1d853648b990412cb223c2e903b
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/265861
Reviewed-by: Semenov, Sergey <sergey.semenov@intel.com>
Tested-by: Semenov, Sergey <sergey.semenov@intel.com>
Reviewed-by: Sotkin, Alexey <alexey.sotkin@intel.com>