[VPlan] Account for a private's original alignment in LLVM-IR vector codegen.
Currently memory that is created for a loop private is lowered by
allocating a continuous chunk of memory for all VF lanes without
accounting for original alignment of the scalar private's memory.
This can lead to buggy vectorized memory access to the private
since codegen uses original alignment of scalar access. This patch
updates lowering of VPAllocatePrivate to account for original memory's
alignment before generating a VF-wide alloca. If the alignments don't
match then we generate VF copies of original alloca (preserving the
original alignment) and build a corresponding vector of pointers.

JIRA : CMPLRLLVM-21649

Change-Id: I4a000fb33322b51e3f1d2626008add15d0f712cc
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/265249
Reviewed-by: Karthik Senthil <karthik.senthil@intel.com>
Tested-by: Karthik Senthil <karthik.senthil@intel.com>
Reviewed-by: Bozhenov, Nikolai <nikolai.bozhenov@intel.com>