[SYCL] Workaround for users with a lot of host accessors.
The issue happens when the user allocates a lot of host accessors to a single
buffer via subbuffers (non-intersecting for write or read/write mode and
regardless of intersections for read-only mode). The culprit of the issue
is that we have a limited number of graph leaves for sake of performance.
When another command is to become a leaf and the limit is to be exhausted we
make the oldest command amongst the leaves to depend on the new one.
This makes host accessors to depend on each other a lot and causes a deadlock
situation.

Change-Id: I94598c0ca3502db4b5552adc0475f0094efba5e4
Signed-off-by: Sergey Kanaev <sergey.kanaev@intel.com>
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/265372
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Romanov, Vlad <vlad.romanov@intel.com>