[OpenCL] fix undefined atomic builtin issue
The first argument of atomic builtins with atomic type is optimized to be
replaced with the one in private address space by InferAddressSpaces pass. But
according to OpenCL spec, the behavior is undefined if the argument with atomic
type is in private address space. So there is no corresponding atomic builtin
defined by OCL CPU RT. Fixed by ignoring this replacement.

JIRA: CMPLRLLVM-22219, CMPLRLLVM-22131

Change-Id: I74e48a7b3c9be0423c4bcb029f7666268ec12482
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/265412
Reviewed-by: Wenju He <wenju.he@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Reviewed-by: Bozhenov, Nikolai <nikolai.bozhenov@intel.com>