[barrier] Fix return value handling in -O0
Functions are not inlined in -cl-opt-disable mode. If return value of a
CallInst crosses barrier and is returned by caller, the value is not
classified as Group B in DataPerValue pass. It is handled in
Barrier::fixReturnValue instead.
However, we still need to mark it as special in markSpecialArguments,
othereise callee will be ignored in fixReturnValue.

This patch also change m_functionToEntryMap type to MapVector so that
print order is deterministic.

Change-Id: Ife482f34ade02b6c1377a21b6b0bab9efcbf2c60
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/265544
Reviewed-by: Wenju He <wenju.he@intel.com>
Tested-by: Wenju He <wenju.he@intel.com>