[VPlan][LLVM IR CG] Allow integer division by safe constant speculation
The main motivation is that it workarounds the problem in CMPLRLLVM-21787 for
which the root cause isn't yet understood.

It also happens to emit an easier to read IR out of vectorizer and most probably
results in more efficient code gen as of now. One drawback is that we lose
information about what lanes aren't needed (i.e. masked away) so it might be
harder for the CG to save some power by masking them off. On the other hand, the
longer term solution for that would probably be a separate intrinsic for the
predicated division (like llvm.vp.sdiv) and not the current predicated
serialization anyway.

Change-Id: I9f79042a99d22786d3ca429a181cbe7b1b1bd5c7
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/265868
Reviewed-by: Elovikov, Andrei <andrei.elovikov@intel.com>
Tested-by: Elovikov, Andrei <andrei.elovikov@intel.com>