CMPLRLLVM-21044: Extensions of aggressive inlining for 519.lbm on Windows. We mark some functions as NoRecurse that are not getting marked that way in the normal function attribute pass. In particular, WeakODR functions are not, because they could be replaced by strong definitions.  But this will not happen if we have "whole program seen". So, we add a traversal that will mark some tiny functions as NoRecurse. We only do tiny functions to save on compile time. We could generalize this for more cases, but right now, this is sufficient for aggressive inlining.
After discussions with Prasad, I have changed the tracking part of this
heuristic to follow the def/use chains to see where the @srcGrid and
@dstGrid pointers are being propagated to. This allows us to prove that
@printf, while getting the dereferenced value of these key pointers, is
not being passed the pointers themselves. This more general algorithm now
works for Windows as well as Linux without special casing.

We also add some LIT tests that mirror the operation of 519.lbm in Windows.
We remove the UNSUPPORTED qualifier because now that the aggressive inlining
code has been generalized, it can recognize both the Linux and Windows code
in 519.lbm for aggressive inlining.

Change-Id: I788bb2212d33594f6dfa0d47f1edff31b8641bd1
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/266397
Reviewed-by: Rivera, Axel Y <axel.y.rivera@intel.com>
Reviewed-by: Battini, Prasad V <prasad.v.battini@intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Tested-by: Cox, Robert <robert.cox@intel.com>
Reviewed-by: Bozhenov, Nikolai <nikolai.bozhenov@intel.com>