[AA][loopopt] Remove ScopedNoAliasAA from loopCarriedAlias
This improves correctness. ScopedNoAlias only asserts "alias()," even
for runtime checks.

The current inclusion of ScopedNoAliasAA in loopCarriedAlias allows
incorrect vectorization.

For example, the following loop with a potentially useful application of
C99 "restrict" should not be vectorized. Removing ScopedNoAlias from
"loopCarriedAlias()" ensures that it is not incorrectly vectorized.

    double Ag[N], Bg[N];
    void foo() {
      for(i=1..N) {
        double* restrict A = i%2 ? Bg : Ag;
        double* restrict B = i%2 ? Ag : Bg;
        A[i] = B[i-1] + 1;
      }
    }

Some of our current performance results (CPU2017's 519 and 538)
mechanically depend on completely eliminating the outer loop dependence
in analogous cases. Fortunately, eliminating the outer loop dependence
in these particular cases is not actually necessary: it is sufficient to
understand only that a particular inner loop has no dependence.

To this end, this change also improves the precision of DDTests.
DDTests' "depends()" has traditionally queried AA only in the context of
the outermost loop to try to eliminate entire dependence edges, meaning
that AA is not queried in the context of an inner loop.

With this change, depends() takes advantage of the fact that "alias()"
may be safe and useful for an inner loop where it was not useful in the
context of the outermost loop. We do this by making "queryAAIndep"
loop-level aware and calling it on behalf of common inner loops if the
entire edge could not be eliminated. That is, for N common loop levels
up to N AA queries will be made instead of just 1.

If AA can break a dependence for an inner loop this is recorded in the
direction vector as an empty ("NONE") direction. Note that "NONE" is
stronger than "EQ", as EQ does not preclude a loop-independent
dependence. A NONE at a particular level allows DDG clients to ignore
DDG edges which have no relationship with an inner loop of interest.

For example, in the loop below it's not safe to use
alias({A, UnknownSize}, {B, UnknownSize}) to completely break the
dependence across i1 because A and B are varying in the outer loop.
However, it can prove that there is no dependence across or within the
i2 loop's iterations. With this change, DDTests will return a direction
vector of (*,) to convey this -- both benefiting from 'restrict' and
correctly preserving outer loop dependencies.

  for(i1=...) {
    float* restrict A = loadA();
    float* restrict B = loadB();
    for(i2=...)
      A[i2] = B[i2-1];
  }

A DDG built on either the i1 or i2 loop will contain (*,) edges between
the two references. However, a DDG client requesting edges for the i2
loop will not see such an edge as it will be filtered out of its view of
the DDG.

If a DDA client for the i1 loop invokes on-demand refinement for i2, the
result will have "isIndependent()" set.

Change-Id: I3a8212e91fcd997081937631ccbe99ef2bd7cb15
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/268176
Reviewed-by: Creech, Timothy M <timothy.m.creech@intel.com>
Tested-by: Creech, Timothy M <timothy.m.creech@intel.com>