[SYCL] Enable L0 online link under envvar
Add temporary support to enable online linking with the Level Zero
driver if the SYCL_ENABLE_LEVEL_ZERO_LINK environment variable is
set.

At the time this code was written, the Level Zero driver online
linking APIs exist, but they don't work.  We think the DPC++ runtime
support is ready, though, and we can enable it via this environment
variable.  This will allow testing with a new driver if one is
available before the next DPC++ release.

Change-Id: I932792e1476d877b1d17e51c4f4f9623d82ed1ed
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/266434
Reviewed-by: Lueck, Gregory M <gregory.m.lueck@intel.com>
Tested-by: Lueck, Gregory M <gregory.m.lueck@intel.com>