[VPO][Paropt] Global Variables not passed as an argument to target/parallel region.
The code extractor does not consider the Global Variables for outlining.
Inner constructs are handled/outlined first, so a private
global variable from outer constructs are seen as globals and hence
not passed as arguments.
Global Vars as arguments are currently handled for target regions.
Support is extended further.
Scenarios where such error was seen:
-parallel firstprivate(@GL) : parallel shared(@GL)
-parallel firstprivate(@GL) : parallel firstprivate(@GL)
-target private(@GL) : teams shared(@GL)
-target private(@GL) : teams firstprivate(@GL)
Added Utilities to call a functor on all Items for required clauses.

JIRA: CMPLRLLVM-21510

Change-Id: I4e94840eb50176727d24aad12f6e5643637402e3
Signed-off-by: Garima Gupta <garima.gupta@intel.com>
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/266640
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/270737