[VPlan][SVA] Avoid out-of-bounds access on parameters list of VectorVariant.
Intel indirect calls are known to have an additional argument
operand which is pointer to call, and this argument is not part
of the corresponding VectorVariant's parameters list. This patch
updates SVA's compute for VPCalls using VectorVariants to account
for this corner case and avoid any out-of-bounds access. SVA nature
of the first argument for such calls is determined using DA.

JIRA : CMPLRLLVM-22230

Change-Id: I984f69defb4b8f69bfb6579ac83cfe74d994f9d4
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/266612
Reviewed-by: Mitropoulou, Konstantina <konstantina.mitropoulou@intel.com>
Reviewed-by: Kozhukhov, Sergey S <sergey.s.kozhukhov@intel.com>
Tested-by: Karthik Senthil <karthik.senthil@intel.com>
Reviewed-by: Bozhenov, Nikolai <nikolai.bozhenov@intel.com>