CMPLRLLVM-22269: Fix to prevent branches with the same target when devirtualizing
The devirtualization process collects the target functions for a particular
indirect call from multiple vtables. It is possible that a target function
is in multiple vtables, which makes repeated entries in the array that
contains the target functions. The issue is that multiversioning process
can generate two branches comparing with the same target. For example,
assume that iCall is a pointer with the address of a function that is the
target function, the devirtualization process can generate something like
this:

   if (iCall == Derived::foo)
     call Derived::foo();
   else if (iCall == Derived2::foo)
     call Derived2::foo();
   else
     call Derived2::foo();

This could happen if Derived2::foo() is in the multiple vtables, e.g.
in the vtable of Derived2 and Derives 3 (assuming that there is another
class called Derived3).

This change set makes sure that if the branch for a target function is
generated then we don't need to make it again. From the example above,
the result will be:

   if (iCall == Derived::foo)
     call Derived::foo();
   else
     call Derived2::foo();

Since the only targets are Derived::foo and Derived2::foo.

Change-Id: I4a28a85af8e5cf04c70ae264ddfee05a08d72ad3
Reviewed-on: https://git-amr-2.devtools.intel.com/gerrit/266802
Reviewed-by: Battini, Prasad V <prasad.v.battini@intel.com>
Reviewed-by: Cox, Robert <robert.cox@intel.com>
Tested-by: xmain_checkin <lab_iclt@iclgerrit.sc.intel.com>
Tested-by: zperf_checkin_xmain <lab_iclz@iclgerrit.sc.intel.com>
Reviewed-by: Bozhenov, Nikolai <nikolai.bozhenov@intel.com>